requirements.txt
google-cloud-storage==2.9.0
google-cloud-bigquery==3.11.0
google-api-python-client==2.92.0